Move replaceEnvVars from Client to YarnSparkHadoopUtil. Minor refactor to share a common variable for the name regex pattern.
